LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-dynamic-expected.txt


FAIL Dynamically changing view-timeline attachment assert_equals: div75 expected "75" but got "0"
FAIL Dynamically changing view-timeline-axis assert_equals: vertical expected "25" but got "-1"
FAIL Dynamically changing view-timeline-inset assert_equals: without inset expected "25" but got "-1"
FAIL Element with scoped view-timeline becoming display:none assert_equals: display:block expected "25" but got "-1"
